python-agent/pyproject.toml
[project]
name = "bfm-agent"
version = "0.1.0"
description = "BFM Medical Copilot Python Agent with GPT 5.2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "openai>=1.60.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "supabase>=2.10.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "pypdf2>=3.0.0",
    "tiktoken>=0.8.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
    "pyyaml>=6.0.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
